refactor: changed column avg_price in portfolio_position scheme type from int to float

+++ backend/app/schemas/portfolio_position.py
@@ -5,7 +5,7 @@ class PortfolioPositionBase(BaseModel):
     portfolio_id: int
     asset_id: int
     quantity: int
-    avg_price: int
+    avg_price: float
 
 class PortfolioPositionCreate(PortfolioPositionBase):
     pass
